+++ CHANGELOG.md
@@ -3,6 +3,13 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 
+[0.X.X] - 2024-XX-XX
+--------------------
+* Enhancements
+  * Added more partial load options for the GNSS line of site data.
+* Maintenance
+  * Updated datetime calls to `utcnow` to be `now` with timezone UTC.
+
 [0.2.0] - 2024-03-15
 --------------------
 * Enhancements